--- client/MEWE/MEWE/MainView/CircleView.swift
@@ -1,28 +1,0 @@
-//
-//  CircleView.swift
-//  MEWE
-//
-//  Created by 1v1 on 2021/01/23.
-//
-
-import SwiftUI
-
-struct CircleView: View {
-    var body: some View {
-        ZStack{
-            Button(action: /*@START_MENU_TOKEN@*/{}/*@END_MENU_TOKEN@*/){
-                Circle()
-                    .fill(Color.orange)
-                    .frame(width: 70, height: 70)
-            }
-            
-            
-        }
-    }
-}
-
-struct CircleView_Previews: PreviewProvider {
-    static var previews: some View {
-        CircleView()
-    }
-}
