+++ client/MEWE/MEWE/RecordEmoji/View/RecordEmojiView.swift
@@ -1,0 +1,27 @@
+//
+//  RecordEmojiView.swift
+//  MEWE
+//
+//  Created by 1v1 on 2021/01/23.
+//
+
+import SwiftUI
+
+
+struct RecordEmojiView: View {
+    var body: some View {
+        Text("Record Emoji View")
+    }
+}
+
+extension RecordEmojiView{
+    private func buttonPressed(with title: String){
+        // if self.selectedTitles.
+    }
+}
+
+struct RecordEmojiView_Previews: PreviewProvider {
+    static var previews: some View {
+        RecordEmojiView()
+    }
+}
