+++ client/MEWE/MEWE/Today/View/TodayEmojiView.swift
@@ -1,5 +1,5 @@
 //
-//  TodayEmoji.swift
+//  TodayEmojiView.swift
 //  MEWE
 //
 //  Created by 1v1 on 2021/01/23.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct TodayEmoji: View {
+struct TodayEmojiView: View {
     var body: some View {
         Text("Today Emoji 영역")
             .frame(width: 320, height: 400)
@@ -17,6 +17,6 @@ struct TodayEmoji: View {
 
 struct TodayEmoji_Previews: PreviewProvider {
     static var previews: some View {
-        TodayEmoji()
+        TodayEmojiView()
     }
 }
