+++ client/MEWE/MEWE/UserProfile/View/UserProfileView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct UserProfileView: View {
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        Text("User profile")
     }
 }
